Recurrent

+++ opennn/config.h
@@ -36,7 +36,7 @@
 
 namespace OpenNN
 {
-    typedef float type;
+    typedef double type;
 }
 
 //#define align 64;
